Changed health to int, added functionality for third potentiometer.
To Fix: Target zones for spells do not clear on death and player cannot change the abilities of their other characters when a character dies.

+++ Unity Project/Comp140 - Unity Project/Assets/Boards/FullBoard.cs
@@ -80,7 +80,7 @@ public class FullBoard : MonoBehaviour
 
     #endregion
 
-    public void Attack(int space, float damage)
+    public void Attack(int space, int damage)
     {
         if (IsSpaceValid(space))
         {
@@ -89,7 +89,7 @@ public class FullBoard : MonoBehaviour
         }
     }
 
-    public void Heal(int space, float heal)
+    public void Heal(int space, int heal)
     {
         if (IsSpaceValid(space))
         {
